Refactor and reorganize test files for dashboard and superAdmin controllers
