+++ ChatHavenApp/backend/src/__tests__/controllers/dashboardControllerMember.test.ts
@@ -1,11 +1,11 @@
 import request from 'supertest';
 import express from 'express';
-import dashboardRoutes from '../routes/dashboardRoutes';
-import authenticate from '../middlewares/authMiddleware';
-import { checkTeamPermission } from '../middlewares/permissionMiddleware';
-import { TeamRole } from '../enums';
-import TestHelpers from './testHelpers';
-import { Role } from '../enums';
+import dashboardRoutes from '../../routes/dashboardRoutes';
+import authenticate from '../../middlewares/authMiddleware';
+import { checkTeamPermission } from '../../middlewares/permissionMiddleware';
+import { TeamRole } from '../../enums';
+import TestHelpers from '../testHelpers';
+import { Role } from '../../enums';
 
 const app = express();
 app.use(express.json());
